working code,added the SIGCONT,SIGSTOP - SIGSTOP is not getting caught

+++ lab1/D.cpp
@@ -1,3 +1,4 @@
+#include<iostream>
 #include<stdio.h>
 #include<sys/wait.h>  // for wait
 #include<signal.h> // for signal
@@ -8,8 +9,14 @@
 #include <sys/types.h> 
 #include <unistd.h> 
 #include<fcntl.h>  // for files reading
-
+using namespace std;
+int val=0;
+void handler(int signum){
+    val=1;
+    printf("interuppt happened");
+}
 int main(){
+    signal(SIGINT,handler);
     char * myfifo="/tmp/myfifo";
     mkfifo(myfifo,0666);
     int pid;
@@ -17,8 +24,11 @@ int main(){
      read(dfd,&pid,1024);
     close(dfd);
     while(1){
-        sleep(5);
-        kill(pid,SIGUSR1);
+        // sleep(5);
+        if(val) {
+            kill(pid,SIGUSR1);
+            val=0;
+        }
     }
     return 0;
 }
